requirements.txt
APScheduler==3.6.3
asgiref==3.9.1
cachetools==4.2.2
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.6
idna==3.10
pillow==11.3.0
python-dotenv==1.1.1
python-telegram-bot==13.15
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
tornado==6.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
